+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 * TypeToStringUtils
     - Add shortcuts (to simplify common usages):
         - toString(Type)
-        - toStringIgnoringVariables(Type)        
+        - toStringIgnoringVariables(Type)    
+* TypeUtils
+    - Fix isAssignableBounds() for proper support of complex wildcards where none of left types is assignable 
+       to all right types            
 * Improve types tracking: tracked types now analyzed for dependent variables to extract all possible type information
 * Fix reversed generic variables declaration support (#3)
 * Fix TypesWalker: processing should not continue after incompatible types detection